Introduce a new `render::OwnedPrimitives` type as an owned alternative to the `render::Primitives` type.
Remove the `image::Map` argument from the Primitives type in favour of
passing it directly to the backend as necessary. As a result, there is
no longer any way for the `image::Map` to notify the `Ui` of when a
redraw is necessary due to some mutation of state. For now the user is
required to manually check whether redraw is required via the
`image::Map`'s `trigger_redraw` field (which will also need to be
manually reset after drawing) until we can come up with a nicer
automated solution.

Added new functions to the piston and piston_draw backends for drawing
individual `Primitive`s.