Add a new `CollapsibleArea` widget.
When "open", the `CollapsibleArea::set` method returns `Some(Area)`.
This `Area` provides a `widget::Id` and a `set` method for positioning a
widget of the user's choice directly under the arrow+text bar.

When "closed" the `CollapsibleArea::set` method returns `None`.

The `CollapsibleArea::set` method also returns an `Option<Event>` which
indicates whether or not the area has just been closed or opened by the
user.

This wdget is useful for giving the user the option to expand or
collapse parts of the user interface. I think this tends to work best
for long columns that contain multiple different areas of settings or
information.