Add new Grid widget and add it to the plot_path example
Using this along with the PlotPath widget should simplify plotting
setups.

There is still room for combining the two into a more user friendly,
high-level widget (along with axis descriptors, text, etc). I'm not sure
if we want to provide that from conrod due to the number of ways this
could be implemented.

This might also be useful placed behind EnvelopEditor too.

Also adds a missing thickness method to the PlotPath widget.