Make the glium backend glyph cache size configurable
Previously the size of the glyph cache was always equal to the size of
the facade's framebuffer dimensions with which it was created. This
could easily cause text glitching problems if the cache size needed for
the glyphs would exceed the size of the original framebuffer dimensions.

This PR allows for manually specifying the dimensions of the
glium::backend::Renderer's inner glyph cache upon construction.