Fix vulkano validation errors by uploading all glyph data at once
Before this PR the method in which we uploaded glyph data did not abide
by the vulkano spec. Here's an example of an error returned by the
standard validation layer:

```
VUID-VkBufferImageCopy-bufferOffset-00194(ERROR / SPEC): msgNum: 0 - vkCmdCopyBufferToImage(): pRegion[0] bufferOffset 0x242d must be a multiple of 4. The Vulkan spec states: bufferOffset must be a multiple of 4 (
https://www.khronos.org/registry/vulkan/specs/1.1-extensions/html/vkspec.html#VUID-VkBufferImageCopy-bufferOffset-00194)
    Objects: 1
        [0] 0x13, type: 10, name: NULL
```

Previously this was occurring each time a glyph was uploaded at an
offset that was not a multiple of 4. This PR fixes this by uploading all
pixel data to the texture at once.