Make delta_xy of Drag events add up to total_delta_xy
With high frame rates, slow mouse movement or high drag thresholds,
Ui::handle_event omits any Drag events whose delta_xy would be below the
threshold. That causes the sum of all delta_xy of sequential Drag events not
to add up to total_delta_xy.

The solution implemented here is to compare total_distance_xy instead of
delta_xy against the threshold. This is admittedly not ideal. A better
solution would be to set a flag when the first movement exceeds the threshold,
then generate Drag events as long as the flag is set. Once the mouse button
is released, reset the flag to stop generating Drag events.