Remove 'build project' instruction from guide
Concerning the guide, chapter 2 (Getting Started), section
"Running the Conrod Examples".

Previously it was suggested that you'd build the entire conrod
project, to check out the examples (the previous step 3).
This commit removes this step and alters the wording of the
goal a bit.

More explanations and the reasoning can be found in the PR.